requirements.txt
# requirements.txt for The Ledger
langgraph>=0.0.10
anthropic>=0.15.0
asyncpg>=0.29.0
pydantic>=2.5.0
fastmcp>=0.1.0
reportlab>=3.6.13
openpyxl>=3.1.2
faker>=20.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
kafka-python>=2.0.2
